flink-rss. fix check-style.
