+++ client-flink/common/pom.xml
@@ -44,7 +44,7 @@
             <groupId>org.apache.flink</groupId>
             <artifactId>flink-runtime</artifactId>
             <scope>provided</scope>
-            <version>1.14.6</version>
+            <version>${flink.version}</version>
         </dependency>
         <dependency>
             <groupId>org.apache.hadoop</groupId>
